Added demo for 16 mics
Works with 8 or 16 mics in DDR.

Uses decimator that processes samples on multiple threads.

Added API for enable/disable assertion (for debug purposes).

Optionally support button input for selection of which two MICs
are output to the I2S DAC.

Port/Pin configuration is done via app_config.h settings.
Set MIC_TILE in app_conf.h to match desired configuration along with
other options.

When using on Tile 0, USB may not be available and push buttons and
LEDs are not to be used (LEDs will indicate MIC Data for 4 of the
8 data lines). For button mic selection functionality is provided
via another pin (X0D12, J12 - Pin2) with an internal pull-up applied.

When using Tile 1, another pin must be used to drive CODEC_RST_N
high (using the test point on the board), in this example, X1D38
(Port 1O) is used for this purpose. A 3k3 series resistor is used on
this pin which is sufficient to override the weak pull-down on the
board while avoiding signal contention.