Simplify app_par_decimator
This example demonstrates creating and using a custom mic array. The
`custom` component, MyTwoStageDecimator, runs two decimator threads in
parallel. This commit simplifies the app so that instead of processing
four microphones per thread (eight in total), it now processes one
microphone per thread (two in total). This allows the example to run on
the L71 board.