Add tests to profile the mic_array's memory and MIPS.
These tests autogenerate the MIPS and memory usage tables that are included in the
documentation.
They also compare the MIPS and memory against committed reference json files and
flag any deviation above a threshold.
Run with pytest --update to update the reference json and rst files in case an
update is required.